further abstract teams cache to its own file
